moves old script + example implementation of abs class
