+++ src/subjugator/gnc/subjugator_centroids/subjugator_centroids/green_tracker.py
@@ -1,78 +1,47 @@
+from subjugator_centroids.centriod_finder import CentriodFinder
 import cv2
+from cv2.typing import MatLike
 import numpy as np
 
-def get_lime_green_centroid(image):
-    """
-    Detect lime green areas in the image and return the centroid (x, y).
-    Returns None if no lime green region is detected.
-    """
-    # Convert BGR to HSV
-    hsv = cv2.cvtColor(image, cv2.COLOR_BGR2HSV)
+# example implementation of centriod abstract class, this one tracks green objects
+class GreenTracker(CentriodFinder):
+    def __init__(self, topic_name: str):
+        self.topic_name_: str = topic_name
 
-    # Define HSV range for lime green
-    lower_green = np.array([40, 100, 100])
-    upper_green = np.array([80, 255, 255])
+    @property
+    def topic_name(self) -> str:
+        return self.topic_name_
 
-    # Threshold the HSV image to get only green colors
-    mask = cv2.inRange(hsv, lower_green, upper_green)
+    def find_centriod(self, frame: MatLike) -> tuple[int, int] | None:
+        """
+        Detect lime green areas in the image and return the centroid (x, y).
+        Returns None if no lime green region is detected.
+        """
+        # Convert BGR to HSV
+        hsv = cv2.cvtColor(frame, cv2.COLOR_BGR2HSV)
 
-    # Find contours in the mask
-    contours, _ = cv2.findContours(mask, cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_SIMPLE)
+        # Define HSV range for lime green
+        lower_green = np.array([40, 100, 100])
+        upper_green = np.array([80, 255, 255])
 
-    if not contours:
-        return None, mask
+        # Threshold the HSV image to get only green colors
+        mask = cv2.inRange(hsv, lower_green, upper_green)
 
-    # Find the largest contour
-    largest = max(contours, key=cv2.contourArea)
+        # Find contours in the mask
+        contours, _ = cv2.findContours(mask, cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_SIMPLE)
 
-    # Compute centroid
-    M = cv2.moments(largest)
-    if M["m00"] == 0:
-        return None, mask
+        if not contours:
+            return None
 
-    cx = int(M["m10"] / M["m00"])
-    cy = int(M["m01"] / M["m00"])
-    return (cx, cy), mask
+        # Find the largest contour
+        largest = max(contours, key=cv2.contourArea)
 
+        # Compute centroid
+        M = cv2.moments(largest)
+        if M["m00"] == 0:
+            return None
 
-def display_debug_image(image, mask, centroid):
-    """
-    Displays the original image, the binary mask, and overlays the centroid (if found).
-    """
-    # Clone the original image
-    display_img = image.copy()
-
-    # Overlay centroid if it exists
-    if centroid:
-        cv2.circle(display_img, centroid, 5, (0, 0, 255), -1)
-        cv2.putText(display_img, f"Centroid: {centroid}", (centroid[0] + 10, centroid[1]),
-                    cv2.FONT_HERSHEY_SIMPLEX, 0.5, (0, 0, 255), 2)
-
-    # Show images
-    cv2.imshow("Lime Green Detection", display_img)
-    cv2.imshow("Mask", mask)
-
-def main():
-    cap = cv2.VideoCapture(0)  # Use your webcam
-
-    if not cap.isOpened():
-        print("Cannot open camera")
-        return
-
-    while True:
-        ret, frame = cap.read()
-        if not ret:
-            print("Failed to grab frame")
-            break
-
-        centroid, mask = get_lime_green_centroid(frame)
-        display_debug_image(frame, mask, centroid)
-
-        if cv2.waitKey(1) & 0xFF == ord('q'):
-            break
-
-    cap.release()
-    cv2.destroyAllWindows()
-
-if __name__ == "__main__":
-    main()
+        cx = int(M["m10"] / M["m00"])
+        cy = int(M["m01"] / M["m00"])
+        return (cx, cy)
+
